Supermart Grocery SalesRetail Analysis/EDA
About Dataset
This is a fictional dataset created for helping the data analysts to practice exploratory data analysis and data visualization. The dataset has data on orders placed by customers on a grocery delivery application.

The dataset is designed with an assumption that the orders are placed by customers living in the state of Tamil Nadu, India.

Please DO NOT reproduce the same dataset without giving me the credits. If you like this dataset, please consider upvoting.